refactor: move ExecutionPlan and related structs into dedicated mod
